Chatbot Nha tuyen dung 2/requirements.txt
flask==2.3.3
flask-cors==4.0.0
langchain==0.0.335
langchain-google-genai==0.0.5
faiss-cpu==1.7.4
sentence-transformers==2.2.2
python-dotenv==1.0.0
pypdf==3.17.1
tiktoken==0.5.1
numpy>=2.2.0,<2.3.0
scikit-learn==1.2.2
google-generativeai==0.3.1